requirements.txt
# Core
pyspark>=3.4,<4.0
pyyaml>=6.0

# Snowflake — the Spark connector jar is loaded via spark.jars.packages;
# the Python snowflake-connector is only needed for standalone scripts.
snowflake-connector-python>=3.6

# PostgreSQL — JDBC driver loaded via spark.jars.packages;
# psycopg2 is useful for admin / health-check scripts.
psycopg2-binary>=2.9

# AWS
boto3>=1.34

# Testing
pytest>=8.0
pytest-cov>=5.0
